requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.14.1
email-validator==2.2.0
certifi==2025.8.3
rfc5424-logging-handler